+++ src/main/java/io/elice/shoppingmall/product/Entity/Review/Request.java
@@ -17,10 +17,10 @@ import lombok.*;
 public class Request extends BaseEntity { //Member Entity 받으면 연관관계 매핑 마무리
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column
-    private Integer memberId;
+    private Long memberId;
 
     @Column(nullable = false)
     private String title;
